requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
openai==1.3.0
python-dotenv==1.0.0
httpx==0.25.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.6
bcrypt==4.2.0
PyJWT==2.8.0
asyncpg==0.30.0
